+++ packages/react-magma-dom/src/components/Table/Table.test.js
@@ -325,7 +325,7 @@ describe('Table', () => {
 
     expect(button).toHaveStyleRule(
       'outline',
-      `2px dotted ${magma.colors.focusInverse}`,
+      `2px solid ${magma.colors.focusInverse}`,
       {
         target: ':focus',
       }